Resolve missing lapack/blas symbols.
On my linux system, I loaded a lapack module, which includes
lapacke and lapack and blas.
But linking openturbine failed with unresolved lapack and blas
symbols until I explicitly added lapack and blas to the
CMakeLists.txt file which was already referencing lapacke.

Apparently this isn't needed on all systems, but perhaps it is
good to explicitly reference the lapack and blas libraries?