Adds a simple mesh data structure for beam elements and a factory function
This data structure provides an association of elements to nodes, so that
we can loop elements and extract which nodes are owned by each one.  State
data (positions, velocities, etc) will be stored externally and reference
this stucture for sizing/access.  The key assembly loops will be parallel
over elements, then gather/scatter based on nodal indicies.

This is a very simple mesh, but should meet our needs for now.  One
requirement is that all elements have the same number of nodes.