Modified Generalized alpha solver to use the field_data and mesh data structures.  Outer loops are now over the nodes.  Several places where BLAS calls could be used now use them for clarity.
One major change is that a new "state" is not created every iteration.
Instead, the existing data is modified in place, which reduces the number
of allocations needed.  To save state data, the user will need to do so
in between iterations manually.

This request has only ported over one unit test to use the new sturcture.
Futher testing using the heavy top problem is underway.