Ported additional unit tests.
This work uncovered a missuse of the KokkosBlas::gemm function.
namely, the "C" matrix cannot be the view as the A or the B for
the function C = a*C + b*A*B.  In this case, C gets modified before
the multiplication happens, thus changing our A (or B) matrix.