Created new interpolation routine for derivatives
The Interpolate routine in solver.h took a jacobian parameter and
applied it to all calcs even when the value was 1. This commit splits
Interpolate into Interpolate and InterpolateDerivative.
Interpolate no longer has a Jacobian argument, but it performs
normalization of the quaternion component of the coordinates.
The InterpolateDerivative take the Jacobian argument and is basically
the same as the previous version of Interpolate. Both functions were
reworked to use KokkosBlas routines.