Changes to allow Cuda build/run
The biggest change it to TangentOperator, which now takes templated
parameters.  This is to allow the first argument to have a strided
layout, which it will when using a subview on device.  There are several
solutions to this problem, but templating on the View type is the
easiest to read.