This commit fixes CI failures when testing on MacOS-latest by pinning to MacOS-13.
MacOS-latest uses arm processors on GitHub.  The setup library we use does not
seem to work on this latest hardware.  Also, the link stage for OpenTurbine fails
in this case because some system libraries that it finds are x86 when it is building
for the native arm.  Sticking to 13 and x86 will let us keep testing on Mac
and improve our trust in CI.  Once MacOS-latest has become better supported, we can look at
introducing it into our process.