Removed ElemIndices data structure in favor of using two Views of size_t.
After recent changes, most of ElemIndices was unused.  Also, in many cases,
only one piece of information was needed in a given location.  This simpler
formation better communicates to the user what information is needed where.

In theory, it seems like it should be a hair faster than the old implementation,
but any differences one way or the other are expected to be unmeasurable.