Added CreateSystem and CreateSystemWithSolver methods to model.
These methods give users a shorter way to create the structures needed
to use OpenTurbine.  CreateSystem just creates the State, Elements, and
Constraints, while CreateSystemWithSolver also creates the Solver.

To support this, a copy constructor was added to Solver.  This is needed
because the default copy constructor was not properly initializing the new
Kernel Handles needed for sparse matrix addition and multiplication.