Rather than using the default algorithm, use KK_DENSE algorithm.
The result is substantial performance improvements (10-30%) for the Floating Platform,
IEA15 Rotor, and IEA15 Rotor with 3000 elements on CPU.  Similar performance
improvements are seen on GPU, with the exception of the IEA15 with 3000 elements.
However, the performance of that test is dominated by the system solve, so the result
is no net change in performance.