Added initial SuperLU-MT solver.
This one will likely need some tuning.  There are a lot of knobs to twist.

It is currently hardcoded to use 1 processor.  This should be generalized.

For the 3000 blade problem, it requires 32 total system solves versus 25 for
other solvers.  This might be rememdied by adding a refinement to the solve (dgsrfs).