Modified kernels to allow execution on host even when Kokkos is configured for CUDA (#407)
This is the culmination of a series of PRs to decouple where we run from the default Kokkos configuration. It modifies all of the execution policies at each kernel launch to be based on the desired execution space as well as all of our scratch views. It adds two regression tests which run on host even when built for CUDA.