Improves the logic in the solver constructor to deduce which DSS to use. (#416)
* Improves the logic in the solver constructor to deduce which DSS to use.

The logic now takes into account the requested execution space for the solver,
selecting a host solver if it is requested, even if OpenTurbine is built with
device support.  The new order of solver preference takes into account recent
benchmarking, which suggest KLU and SuperLU give the best results on CPU.  On
CUDA, cuDss gives the fastest results, but cuSolver_Sp is more robust at this time.

* clang-format