Change cuDSS to use more accurate reordering algorithm (#430)
This is necessary to prevent NaNs on the full turbine configuration

This does significantly reduce performance for the IEA15Rotor problem with 3000 blades, but the extra stability and accuracy is necessary. For that problem, cuSolverSP now presents the most performant GPU-based solver, while cuDSS is about twice as slow, and copying to Host and using KLU is twice as slow as that.

For the 3 blade problem, performance with the new algorithm selection is about the same as the default.

For more details, see: https://docs.nvidia.com/cuda/cudss/types.html#cudssconfigparam-t