Changes Rotor and Rotating beam tests to use model's blade manipulation functions
This is more in line with how we expect users to interact with Model, so we should
demonstrate and test it.  It also adds a regression check to the IEA15Rotor test
which ensured that the use of this API is the same as applying the transformation
to the nodes directly.